bpf: Avoid implicit shorten-64-to-32 in clang 19
```
bpf_nat_tests.c:1141:10: error: implicit conversion loses integer precision: 'long' to '__u32' (aka 'unsigned int') [-Werror,-Wshorten-64-to-32]
 1141 |         iters = loop(SNAT_TEST_ITERATIONS, snat_callback_tcp, &cb_ctx, 0);
      |               ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
bpf_nat_tests.c:1272:10: error: implicit conversion loses integer precision: 'long' to '__u32' (aka 'unsigned int') [-Werror,-Wshorten-64-to-32]
 1272 |         iters = loop(SNAT_TEST_ITERATIONS, snat_callback_udp, &cb_ctx, 0);
      |               ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
```

Relates: https://github.com/cilium/cilium/pull/36186

+++ bpf/tests/bpf_nat_tests.c
@@ -1131,7 +1131,7 @@ int test_nat4_port_allocation_tcp_check(struct __ctx_buff *ctx)
 	struct snat_callback_ctx cb_ctx = {
 		.ctx = ctx,
 	};
-	__u32 iters;
+	long iters;
 
 	test_init();
 	/* This test checks the effectiveness of port allocation algorithm in SNAT.
@@ -1262,7 +1262,7 @@ int test_nat4_port_allocation_udp_check(struct __ctx_buff *ctx)
 	struct snat_callback_ctx cb_ctx = {
 		.ctx = ctx,
 	};
-	__u32 iters;
+	long iters;
 
 	test_init();
 	/* This test checks the effectiveness of port allocation algorithm in SNAT.
